Moved DsstneContext to amazon/dsstne/runtime and added Makefile target to produce a shared library for engine
