+++ src/amazon/dsstne/runtime/Makefile
@@ -7,10 +7,10 @@ include ../Makefile.inc
 SOURCES_DIR := .
 
 HEADERS := $(wildcard *.h *.hpp)
-HEADERS_BUILD_DIR := $(BUILD_DIR)/include/amazon/dsstne/engine
+HEADERS_BUILD_DIR := $(BUILD_DIR)/include/amazon/dsstne/runtime
 
-OBJS_BUILD_DIR := $(BUILD_DIR)/tmp/engine/cpp
-CU_OBJS_BUILD_DIR := $(BUILD_DIR)/tmp/engine/cuda
+OBJS_BUILD_DIR := $(BUILD_DIR)/tmp/runtime/cpp
+CU_OBJS_BUILD_DIR := $(BUILD_DIR)/tmp/runtime/cuda
 $(shell mkdir -p $(OBJS_BUILD_DIR))
 $(shell mkdir -p $(CU_OBJS_BUILD_DIR))
 
@@ -20,38 +20,39 @@ CU_SOURCES=$(wildcard *.cu)
 OBJECTS := $(SOURCES:%.cpp=$(OBJS_BUILD_DIR)/%.o)
 CU_OBJECTS := $(CU_SOURCES:%.cu=$(CU_OBJS_BUILD_DIR)/%.o)
 
+LIB_DSSTNE :=$(BUILD_DIR)/lib/libdsstne.a
+
+LDFLAGS ?= -fPIC
+#LDFLAGS ?= -fPIC -Wl,--whole-archive $(LIB_DSSTNE) -Wl,--no-whole-archive
+
+LIBS = \
+    $(CU_LIBS) \
+    -L$(BUILD_DIR)/lib
+
+LOAD_LIBS = \
+    $(CU_LOADLIBS) \
+    -ldsstne_engine
+
 DEP := $(OBJECTS:.o=.d)
 CU_DEP := $(CU_OBJECTS:.o=.d)
 
-CU_INCLUDES += -I../utils
-
 LIB_BUILD_DIR := $(BUILD_DIR)/lib
 
-all: $(LIB_BUILD_DIR)/libdsstne.a $(LIB_BUILD_DIR)/libdsstne_engine.so
+all: $(LIB_BUILD_DIR)/libdsstne_runtime.so
 
-$(LIB_BUILD_DIR)/libdsstne.a: $(OBJECTS) $(CU_OBJECTS)
-	$(info ========== Creating libdsstne.a ==========)
-	mkdir -p $(LIB_BUILD_DIR)
-	ar -cvr $@ $^
-	$(info ========== Copying amazon/dsstne/engine headers ==========)
+$(LIB_BUILD_DIR)/libdsstne_runtime.so: $(OBJECTS) $(CU_OBJECTS)
+	$(info ========== Creating libdsstne_runtime.so ==========)
+	mkdir -p $(BUILD_DIR)/lib
+	$(CC) -shared $(LDFLAGS) $(LIBS) $(OBJECTS) $(CU_OBJECTS) -o $@ $(LOAD_LIBS)
+	$(info ========== Copying amazon/dsstne/runtime headers ==========)
 	mkdir -p $(HEADERS_BUILD_DIR)
 	cp $(HEADERS) $(HEADERS_BUILD_DIR)
 
-$(LIB_BUILD_DIR)/libdsstne_engine.so: $(OBJECTS) $(CU_OBJECTS)
-	$(info ========== Creating libdsstne_engine.so ==========)
-	mkdir -p $(BUILD_DIR)/lib
-	$(CC) -shared $(LDFLAGS) $(CU_LIBS) $(OBJECTS) $(CU_OBJECTS) -o $@ $(CU_LOADLIBS)
-
 clean:
-	rm -f *cudafe* *.fatbin.* *.fatbin *.ii *.cubin *cu.cpp *.ptx *.cpp?.* *.hash *.o *.d work.pc*
-	rm -rf $(OBJS_BUILD_DIR) $(CU_OBJS_BUILD_DIR) $(HEADERS_BUILD_DIR) $(LIB_BUILD_DIR)/libdsstne.a
+	rm -rf $(OBJS_BUILD_DIR) $(CU_OBJS_BUILD_DIR) $(HEADERS_BUILD_DIR) $(LIB_BUILD_DIR)/libdsstne_runtime.so
 
 distclean:
-	rm -f *cudafe* *.fatbin.* *.fatbin *.ii *.cubin *cu.cpp *.ptx *.cpp?.* *.hash *.o *.d work.pc*
 	rm -rf $(OBJS_BUILD_DIR) $(CU_OBJS_BUILD_DIR)
-
-# The following statement keeps the .f90 files from getting clobbered; this
-# would happen because they are intermediates.
 
 # Pattern rule for compiling C++
 $(OBJS_BUILD_DIR)/%.o: $(SOURCES_DIR)/%.cpp
